Move code into more sensible place
